+++ src/weapons.c
@@ -1,6 +1,6 @@
 /*
  * weapons.c
- * Copyright (C) 2009-2018 Joachim de Groot <jdegroot@web.de>
+ * Copyright (C) 2009-2025 Joachim de Groot <jdegroot@web.de>
  *
  * NLarn is free software: you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -46,30 +46,31 @@ const char *ammo_class_name[AMMO_MAX] =
 const weapon_data weapons[WT_MAX] =
 {
     /* weapon_t         name                         short name,       class          ammo      dam  acc  material    we     pr tw un ar ob th*/
-    { WT_ODAGGER,       "orcish dagger",             "orc dagger",      WC_MELEE,     AMMO_MAX,  2,   3, IM_IRON,   850,     6, 0, 0, 1, 0, 1 },
-    { WT_DAGGER,        "dagger",                    "dagger",          WC_MELEE,     AMMO_MAX,  3,   5, IM_IRON,   600,    10, 0, 0, 1, 1, 1 },
-    { WT_SLING,         "sling",                     "sling",          WC_RANGED,   AMMO_SLING,  1,   3, IM_CLOTH,  200,    15, 1, 0, 1, 1, 0 },
-    { WT_OSHORTSWORD,   "orcish short sword",        "orc short sword", WC_MELEE,     AMMO_MAX,  5,   2, IM_IRON,  1350,    30, 0, 0, 1, 0, 0 },
-    { WT_SHORTSWORD,    "short sword",               "short sword",     WC_MELEE,     AMMO_MAX,  6,   3, IM_IRON,  1000,    60, 0, 0, 1, 1, 0 },
-    { WT_ESHORTSWORD,   "elven short sword",         "elf short sword", WC_MELEE,     AMMO_MAX,  7,   4, IM_WOOD,  1180,    68, 0, 0, 1, 0, 0 },
-    { WT_OSPEAR,        "orcish spear",              "orc spear",       WC_MELEE,     AMMO_MAX,  8,   2, IM_IRON,  2100,    75, 0, 0, 1, 0, 1 },
-    { WT_SPEAR,         "spear",                     "spear",           WC_MELEE,     AMMO_MAX, 10,   3, IM_IRON,  1800,   120, 0, 0, 1, 1, 1 },
-    { WT_ESPEAR,        "elven spear",               "elf spear",       WC_MELEE,     AMMO_MAX, 11,   4, IM_WOOD,  1600,   140, 0, 0, 1, 0, 1 },
-    { WT_BOW,           "bow",                       "bow",            WC_RANGED,     AMMO_BOW,  3,   4, IM_WOOD,  1000,   220, 1, 0, 1, 1, 0 },
-    { WT_CLUB,          "heavy club",                "club",            WC_MELEE,     AMMO_MAX, 10,   2, IM_WOOD,  3500,    50, 0, 0, 1, 0, 0 },
-    { WT_MACE,          "mace",                      "mace",            WC_MELEE,     AMMO_MAX, 12,   3, IM_IRON,  2600,   160, 0, 0, 1, 1, 0 },
-    { WT_FLAIL,         "flail",                     "flail",           WC_MELEE,     AMMO_MAX, 14,   2, IM_IRON,  2900,   195, 1, 0, 1, 1, 0 },
-    { WT_BATTLEAXE,     "battle-axe",                "battle-axe",      WC_MELEE,     AMMO_MAX, 18,   2, IM_IRON,  2700,   350, 1, 0, 1, 1, 0 },
-    { WT_CROSSBOW,      "crossbow",                  "crossbow",       WC_RANGED, AMMO_CROSSBOW, 5,   3, IM_WOOD,  3500,   600, 1, 0, 1, 0, 0 },
-    { WT_LONGSWORD,     "longsword",                 "longsword",       WC_MELEE,     AMMO_MAX, 22,   3, IM_IRON,  1950,   550, 0, 0, 1, 1, 0 },
-    { WT_2SWORD,        "two-handed sword",          "2-handed sword",  WC_MELEE,     AMMO_MAX, 26,   4, IM_IRON,  3600,  1000, 1, 0, 1, 1, 0 },
-    { WT_SWORDSLASHING, "sword of slashing",         "slashing",        WC_MELEE,     AMMO_MAX, 30,   5, IM_STEEL, 2200,  6000, 0, 0, 1, 0, 0 },
+    { WT_ODAGGER,       "orcish dagger",             "orc dagger",      WC_MELEE,     AMMO_MAX,  2,   3, IM_IRON,     850,     6, 0, 0, 1, 0, 1 },
+    { WT_DAGGER,        "dagger",                    "dagger",          WC_MELEE,     AMMO_MAX,  3,   5, IM_IRON,     600,    10, 0, 0, 1, 1, 1 },
+    { WT_SLING,         "sling",                     "sling",          WC_RANGED,   AMMO_SLING,  1,   3, IM_CLOTH,    200,    15, 1, 0, 1, 1, 0 },
+    { WT_OSHORTSWORD,   "orcish short sword",        "orc short sword", WC_MELEE,     AMMO_MAX,  5,   2, IM_IRON,    1200,    30, 0, 0, 1, 0, 0 },
+    { WT_SHORTSWORD,    "short sword",               "short sword",     WC_MELEE,     AMMO_MAX,  6,   3, IM_IRON,     900,    60, 0, 0, 1, 1, 0 },
+    { WT_ESHORTSWORD,   "elven short sword",         "elf short sword", WC_MELEE,     AMMO_MAX,  9,   4, IM_MITHRIL,  750,   100, 0, 0, 1, 0, 0 },
+    { WT_OSPEAR,        "orcish spear",              "orc spear",       WC_MELEE,     AMMO_MAX,  8,   2, IM_WOOD,    2100,    75, 0, 0, 1, 0, 1 },
+    { WT_SPEAR,         "spear",                     "spear",           WC_MELEE,     AMMO_MAX, 10,   3, IM_WOOD,    1800,   120, 0, 0, 1, 1, 1 },
+    { WT_ESPEAR,        "elven spear",               "elf spear",       WC_MELEE,     AMMO_MAX, 11,   4, IM_WOOD,    1600,   140, 0, 0, 1, 0, 1 },
+    { WT_BOW,           "bow",                       "bow",            WC_RANGED,     AMMO_BOW,  3,   4, IM_WOOD,    1000,   220, 1, 0, 1, 1, 0 },
+    { WT_CLUB,          "heavy club",                "club",            WC_MELEE,     AMMO_MAX, 10,   2, IM_WOOD,    3500,    50, 0, 0, 1, 0, 0 },
+    { WT_MACE,          "mace",                      "mace",            WC_MELEE,     AMMO_MAX, 12,   3, IM_IRON,    2600,   160, 0, 0, 1, 1, 0 },
+    { WT_FLAIL,         "flail",                     "flail",           WC_MELEE,     AMMO_MAX, 14,   2, IM_WOOD,    2900,   195, 1, 0, 1, 1, 0 },
+    { WT_BATTLEAXE,     "battle-axe",                "battle-axe",      WC_MELEE,     AMMO_MAX, 18,   2, IM_IRON,    2700,   350, 1, 0, 1, 1, 0 },
+    { WT_CROSSBOW,      "crossbow",                  "crossbow",       WC_RANGED, AMMO_CROSSBOW, 5,   3, IM_WOOD,    3500,   600, 1, 0, 1, 0, 0 },
+    { WT_LONGSWORD,     "longsword",                 "longsword",       WC_MELEE,     AMMO_MAX, 22,   3, IM_IRON,    1950,   550, 0, 0, 1, 1, 0 },
+    { WT_ELONGSWORD,    "elven longsword",           "elf longsword",   WC_MELEE,     AMMO_MAX, 24,   4, IM_MITHRIL, 1600,   900, 0, 0, 1, 0, 0 },
+    { WT_2SWORD,        "two-handed sword",          "2-handed sword",  WC_MELEE,     AMMO_MAX, 26,   4, IM_IRON,    3600,  1000, 1, 0, 1, 1, 0 },
+    { WT_SWORDSLASHING, "sword of slashing",         "slashing",        WC_MELEE,     AMMO_MAX, 30,   5, IM_STEEL,   2200,  6000, 0, 0, 1, 0, 0 },
     /* unique weapons */
-    { WT_LANCEOFDEATH,  "lance of death",            "lance of death",  WC_MELEE,     AMMO_MAX, 20,   3, IM_WOOD,  2900, 65000, 1, 1, 1, 1, 0 },
-    { WT_VORPALBLADE,   "Vorpal blade",              "Vorpal blade",    WC_MELEE,     AMMO_MAX, 22,   3, IM_STEEL, 1600,  3800, 0, 1, 1, 0, 0 },
-    { WT_SLAYER,        "Slayer",                    "Slayer",          WC_MELEE,     AMMO_MAX, 30,   5, IM_STEEL, 1800,  6800, 0, 1, 0, 0, 0 },
-    { WT_SUNSWORD,      "Sunsword",                  "Sunsword",        WC_MELEE,     AMMO_MAX, 32,   6, IM_STEEL, 1800,  7000, 0, 1, 1, 0, 0 },
-    { WT_BESSMAN,       "Bessman's flailing hammer", "Bessman's",       WC_MELEE,     AMMO_MAX, 35,   6, IM_STEEL, 5800, 10000, 1, 1, 0, 0, 0 },
+    { WT_LANCEOFDEATH,  "lance of death",            "lance of death",  WC_MELEE,     AMMO_MAX, 20,   3, IM_WOOD,    2900, 65000, 1, 1, 1, 1, 0 },
+    { WT_VORPALBLADE,   "Vorpal blade",              "Vorpal blade",    WC_MELEE,     AMMO_MAX, 22,   3, IM_STEEL,   1600,  3800, 0, 1, 1, 0, 0 },
+    { WT_SLAYER,        "Slayer",                    "Slayer",          WC_MELEE,     AMMO_MAX, 30,   5, IM_STEEL,   1800,  6800, 0, 1, 0, 0, 0 },
+    { WT_SUNSWORD,      "Sunsword",                  "Sunsword",        WC_MELEE,     AMMO_MAX, 32,   6, IM_STEEL,   1800,  7000, 0, 1, 1, 0, 0 },
+    { WT_BESSMAN,       "Bessman's flailing hammer", "Bessman's",       WC_MELEE,     AMMO_MAX, 35,   6, IM_STEEL,   5800, 10000, 1, 1, 0, 0, 0 },
 };
 
 /* static functions */
@@ -79,32 +80,6 @@ gboolean weapon_ammo_drop(map *m, item *ammo, const GList *traj);
 static gboolean weapon_pos_hit(const GList *traj,
         const damage_originator *damo,
         gpointer data1, gpointer data2);
-
-int weapon_calc_to_hit(struct player *p, struct _monster *m, item *weapon, item *ammo)
-{
-    g_assert (p != NULL && m != NULL);
-
-    const int to_hit = p->level
-                       + max(0, player_get_dex(p) - 12)
-                       + (weapon ? weapon_acc(weapon) : 0)
-                       + (ammo ? ammo_accuracy(ammo) : 0)
-                       + (player_get_speed(p) / 25)
-                       /* the rule below gives a -3 for tiny monsters and a +4
-                          for gargantuan monsters */
-                       + ((monster_size(m) - MEDIUM) / 25)
-                       - monster_ac(m)
-                       - (monster_speed(m) / 25)
-                       - (!monster_in_sight(m) ? 5 : 0);
-
-    if (to_hit < 1)
-        return 0;
-
-    if (to_hit >= 20)
-        return 100;
-
-    /* roll the dice */
-    return (5 * to_hit);
-}
 
 int weapon_fire(struct player *p)
 {
@@ -448,7 +423,7 @@ static gboolean weapon_pos_hit(const GList *traj,
     {
         /* there is a monster at the position */
         /* the bullet might have hit the monster */
-        if (chance(weapon_calc_to_hit(nlarn->p, m, weapon, ammo)))
+        if (chance(combat_chance_player_to_monster_hit(nlarn->p, m, TRUE)))
         {
             /* hit */
             damage *dam = weapon_get_ranged_damage(nlarn->p, weapon, ammo);
@@ -486,4 +461,36 @@ static gboolean weapon_pos_hit(const GList *traj,
 
     g_free(adesc);
     return retval;
+}
+
+int weapon_instakill_chance(weapon_t wt, monster_t mt)
+{
+    int percentage = 0;
+
+    switch (wt)
+    {
+        /* Vorpal Blade */
+    case WT_VORPALBLADE:
+        if (monster_type_flags(mt, HEAD) && !monster_type_flags(mt, NOBEHEAD))
+            percentage = 5;
+        break;
+
+        /* Lance of Death */
+    case WT_LANCEOFDEATH:
+        /* the lance is pretty deadly for non-demons */
+        if (!monster_type_flags(mt, DEMON))
+            percentage = 100;
+        break;
+
+        /* Slayer */
+    case WT_SLAYER:
+        if (monster_type_flags(mt, DEMON))
+            percentage = 100;
+        break;
+
+    default:
+        break;
+    }
+
+    return percentage;
 }