+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ### Changes:
 * Switch to 256 colours
+* New weapon: the elven longsword - elves may be equipped with it
+* Overhaul the material of some weapon:
+  * All spears are made of wood
+  * The elven short sword is now made of mithril
+  * The flail is now made of wood
+* Reduce the probability of weapons breaking
 * Show carried gold in yellow (suggested by jv84)
 * The potion of see invisible now cures blindness (suggested by jv84)
 * Modified potions of instant healing: they now always cure poisoning